bridge/requirements.txt
sseclient-py
kafka-python
requests